requirements/requirements.txt
Django==1.11
djangorestframework==3.6.2
django-filter==1.0.2
django-extensions==1.7.8
django-cors-headers==2.3.0
PyYAML==3.12
django-allauth==0.31.0
django-rest-auth[with_social]==0.9.1
coreapi==2.3.0
Pillow==4.1.0
django-rest-swagger==2.1.2
gunicorn==19.7.1
psycopg2==2.7.1
click==6.7
python-ldap==3.1.0
django-auth-ldap==1.7.0
django-mama-cas==2.4.0
python-memcached==1.59